Update dependabot. Clean-up kv mode code.
