+++ src/components/GeneralConfiguration.tsx
@@ -5,36 +5,10 @@ import { useLocalStorage } from "~hooks/storage";
 import { defaultOptions } from "~utils/options";
 import { matchPatternsChanged } from "~utils/match-pattern";
 import { setLocalStorage, type MatchPatternError } from "~utils/storage";
-import { parseKeyValuePairs } from "~utils/string";
-import { consoleProxy } from "~utils/logging";
+import { colonSeparatedStringsToMap, mapToColonSeparatedStrings } from "~utils/string";
 
 type GeneralConfigurationProps = {
     enabled: boolean
-}
-
-const colonSeparatedStringsToMap = (strings: string[]): Map<string, string> => {
-    const map = new Map<string, string>()
-    strings.forEach((string) => {
-        const [kvs, _] = parseKeyValuePairs(string, ',', false)
-        consoleProxy.debug('parsed kvs', kvs, 'for string', string)
-
-        kvs.forEach((value, key) => {
-            map.set(key, value)
-        })
-    })
-    consoleProxy.debug('converted strings to map', map)
-    return map
-}
-
-const mapToColonSeparatedStrings = (map: Map<string, string>): string[] => {
-    const strings = []
-    if (map) {
-        map.forEach((value, key) => {
-            strings.push(`${key}:${value}`)
-        })
-    }
-    consoleProxy.debug('converted map to strings', strings, 'from map', map)
-    return strings
 }
 
 const patternErrorsToPills = (patterns: string[], errors: MatchPatternError[]): Map<number, string> => {
@@ -120,9 +94,8 @@ export default function GeneralConfiguration({ enabled }: GeneralConfigurationPr
                 <TagsInput
                     value={attributesStrings}
                     onValueRemoved={(index) => {
-                        const newAttributes = [...attributesStrings]
-                        newAttributes.splice(index, 1)
-                        setLocalStorage({ attributes: colonSeparatedStringsToMap(newAttributes) })
+                        attributesStrings.splice(index, 1)
+                        setLocalStorage({ attributes: colonSeparatedStringsToMap(attributesStrings) })
                     }}
                     onValueAdded={(value) => {
                         attributesStrings.push(value)
@@ -133,16 +106,15 @@ export default function GeneralConfiguration({ enabled }: GeneralConfigurationPr
                     description="Attach additional attributes on all exported logs/traces."
                     placeholder={attributesStrings.length == 0 ? 'key:value, key2:value2' : ''}
                     delimiter={","}
+                    keyValueMode={true}
                 />
                 <TagsInput
                     value={headersStrings}
                     onValueRemoved={(index) => {
                         headersStrings.splice(index, 1)
-                        console.log('value removed at index', index, 'from', headersStrings)
                         setLocalStorage({ headers: colonSeparatedStringsToMap(headersStrings) })
                     }}
                     onValueAdded={(value) => {
-                        consoleProxy.debug('adding header', value)
                         headersStrings.push(value)
                         setLocalStorage({ headers: colonSeparatedStringsToMap(headersStrings) })
                     }}
@@ -151,6 +123,7 @@ export default function GeneralConfiguration({ enabled }: GeneralConfigurationPr
                     description="Include additional HTTP headers on all export requests."
                     placeholder={headersStrings.length == 0 ? 'key:value, key2:value2' : ''}
                     delimiter={","}
+                    keyValueMode={true}
                 />
             </Group>
         </Fieldset>
